requirements.txt
# GLC Platform - Green Lending & Compliance Framework
# Core Dependencies

# Web Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.6
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.45

# NLP & ML
sentence-transformers==5.2.0
transformers==4.57.3
torch==2.9.1

# Vector Search  
faiss-cpu==1.13.2

# PDF Processing
pdfminer.six==20251230
PyPDF2==3.0.1
python-docx==1.2.0

# PDF Report Generation
weasyprint==67.0

# Utilities
numpy==2.4.0
python-jose==3.5.0
passlib==1.7.4
aiofiles==25.1.0

# Development
pytest==9.0.2
httpx==0.28.1

pydantic[email]